+++ menu-api/Repositories/MenuItem_IngredientRepository.cs
@@ -3,7 +3,7 @@ using menu_api.Models;
 
 namespace menu_api.Repositories
 {
-    public class MenuItem_IngredientRepository : IMenuItem_IngredientRepository, IDisposable
+    public class MenuItem_IngredientRepository : IDisposable, IMenuItem_IngredientRepository
     {
         private MenuContext context;
         public MenuItem_IngredientRepository(MenuContext context)
@@ -12,27 +12,31 @@ namespace menu_api.Repositories
         }
 
         //Add ingredient to menuItem
-        public void AddIngredient(MenuItem_Ingredient menuItem_Ingredient)
+        public async Task AddIngredient(MenuItem_Ingredient menuItem_Ingredient)
         {
-            context.MenuItem_Ingredients.Add(menuItem_Ingredient);
-            context.SaveChanges();
+            await context.MenuItem_Ingredients.AddAsync(menuItem_Ingredient);
+            await context.SaveChangesAsync();
         }
 
         //remove ingredient from menuItem
-        public void RemoveIngredient(Guid menuItemId, Guid ingredientId)
+        public async Task RemoveIngredient(Guid menuItemId, Guid ingredientId)
         {
-            MenuItem_Ingredient menuItem_Ingredient = context.MenuItem_Ingredients
-                .Find(menuItemId, ingredientId);
-            context.MenuItem_Ingredients.Remove(menuItem_Ingredient);
-            context.SaveChanges();
+            MenuItem_Ingredient? menuItem_Ingredient = await context.MenuItem_Ingredients
+                .FindAsync(menuItemId, ingredientId);
+            if (menuItem_Ingredient != null)
+            {
+                context.MenuItem_Ingredients.Remove(menuItem_Ingredient);
+                await context.SaveChangesAsync();
+            }
         }
 
         //remove all ingredients from menuItem
-        public void RemoveIngredients(Guid menuItemId)
+        public async Task RemoveIngredients(Guid menuItemId)
         {
-            IEnumerable<MenuItem_Ingredient> menuItem_Ingredients = context.MenuItem_Ingredients.Where(x => x.MenuItemId == menuItemId);
+            IEnumerable<MenuItem_Ingredient> menuItem_Ingredients = context.MenuItem_Ingredients
+                .Where(x => x.MenuItemId == menuItemId);
             context.MenuItem_Ingredients.RemoveRange(menuItem_Ingredients);
-            context.SaveChanges();
+            await context.SaveChangesAsync();
         }
 
         private bool disposed = false;
